requirements.txt
pyserial
pyubx2
matplotlib
numpy
pandas
simplekml
pymap3d
cssrlib
